+++ src/main/java/de/tudarmstadt/informatik/hostage/protocol/cifs/FileInject.java
@@ -30,7 +30,7 @@ public class FileInject {
 
     private Listener fListener;
     SharedPreferences fpref;
-    private int attack_id;
+    private long attack_id;
     private String externalIP;
     private String BSSID;
     private String SSID;
@@ -79,8 +79,8 @@ public class FileInject {
 
     private synchronized void getAndIncrementAttackID(SharedPreferences pref) {
         SharedPreferences.Editor editor = pref.edit();
-        attack_id = pref.getInt("ATTACK_ID_COUNTER", 0);
-        editor.putInt("ATTACK_ID_COUNTER", attack_id + 1);
+        attack_id = pref.getLong("ATTACK_ID_COUNTER", 0);
+        editor.putLong("ATTACK_ID_COUNTER", attack_id + 1);
         editor.commit();
     }
